Refactor module structure: rename 'src' to 'solidspy' and update import paths; update project name and version in pyproject.toml
