+++ src/solidspy/blochutil.py
