+++ src/solidspy/solutil.py
