+++ tests/test_femutil.py
@@ -5,7 +5,7 @@ Test cases for functions on ``femutil`` module
 """
 import numpy as np
 import pytest
-import src.femutil as fem
+import solidspy.femutil as fem
 
 
 #%% Tests for Shape functions and derivatives
