+++ tests/test_uelutil.py
@@ -4,7 +4,7 @@ Test cases for functions on ``uelutil`` module
 
 """
 import numpy as np
-import src.uelutil as uel
+import solidspy.uelutil as uel
 
 
 #%% Continuum elements
